backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
pandas==2.0.3
SQLAlchemy==2.0.20
python-dotenv==1.0.0
